// gestimowebfront-end/src/gs-api/src/models/agence-response-dto.ts
/* tslint:disable */
export interface AgenceResponseDto {
  id?: number;
  idAgence?: number;
  nomAgence?: string;
  telAgence?: string;
  compteContribuable?: string;
  capital?: number;
  emailAgence?: string;
  mobileAgence?: string;
  mobileAgenceSecondaire?: string;
  regimeFiscaleAgence?: string;
  faxAgence?: string;
  sigleAgence?: string;
  adresseAgence?: string;
  boitePostaleAgence?: string;
  profileAgenceUrl?: string;
}